+++ src/index.ts
@@ -1,8 +1,8 @@
-export { PoemType } from "./types/poem-type";
-export { PoemStatus, Poem } from "./types/poem";
+export { PieceType } from "./types/piece-type";
+export { PieceStatus, Piece } from "./types/piece";
 export { PaginationInput } from "./types/pagination";
 export { OperationStatus } from "./types/operationStatus";
-export { StanzaStatus, Stanza } from "./types/stanza";
+export { PartStatus, Part } from "./types/part";
 export { CommonGenre } from "./types/genre";
 export { User, UserStatus } from "./types/user";
 export { userSchema, USER_ERRORS } from "./validation/userSchema";